+++ Pandora/Paradigm/Structure/Modification/Nonempty.hs
@@ -1,4 +1,9 @@
 module Pandora.Paradigm.Structure.Modification.Nonempty where
 
--- | Type synonymous for at least one element data structure
-type family Nonempty (s :: * -> *) = (r :: * -> *) | r -> s
+import Pandora.Paradigm.Schemes (type (<::>))
+import Pandora.Paradigm.Primary.Functor.Maybe (Maybe)
+import Pandora.Paradigm.Primary.Transformer.Construction (Construction)
+
+-- | At least one element data structure
+type family Nonempty (structure :: * -> *) where
+	Nonempty (Maybe <::> Construction t) = Construction t